pyproject.toml
[project]
name = "ytrag-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typesense>=1.3.0",
]
